[BUGFIX] fix printing the OperationLogger to IO
Previously `puts operation.log` would print just an empty line. This is
because `puts` invokes `operation.log.to_s(STDOUT)` that needs to insert
contents into the STDOUT IO.

Now we just overload `to_s` so that if passed an IO object it will write
the operation log into it.